modules: Add the path where we install the helper PAM modules to the .pc file
A test needs to often know where are the helper pam modules
(pam_get_items.so and pam_set_items.so). This patch adds a new variable
to the pkg-config file that lets the user of pam_wrapper query the path
like this:
    $ pkg-config --variable=modules pam_wrapper
    /usr/lib64/pam_wrapper

And use that variable to define their own service files.